process-isolate schedulers: use a unique folder + refac
